+++ sample/src/main/java/ai/devrev/sdk/sample/IdentificationFragment.kt
@@ -1,7 +1,5 @@
 package ai.devrev.sdk.sample
 
-import ai.devrev.sdk.model.Identity
-import ai.devrev.sdk.model.UserInfo
 import ai.devrev.sdk.sample.viewmodel.IdentificationViewModel
 import ai.devrev.sdk.sample.viewmodel.SharedViewModel
 import android.os.Bundle
@@ -30,17 +28,14 @@ class IdentificationFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
+        val isUserIdentifiedCheckbox: CheckBox = view.findViewById(R.id.is_user_identified)
         val identifyUnverifiedUserButton: Button = view.findViewById(R.id.unverified_user_button)
         val identifyVerifiedUserButton: Button = view.findViewById(R.id.verified_user_button)
-        val isUserIdentifiedCheckbox: CheckBox = view.findViewById(R.id.is_user_identified)
         val logOutButton: Button = view.findViewById(R.id.logout_button)
-        val sessionTokenText: EditText = view.findViewById(R.id.session_token_text)
         val unverifiedUserIdText: EditText = view.findViewById(R.id.user_id_text)
-        val updateUserButton: Button = view.findViewById(R.id.update_user_button)
-        val updateUserEmail: EditText = view.findViewById(R.id.update_user_email)
         val verifiedUserIdText: EditText = view.findViewById(R.id.verified_user_id_text)
+        val sessionTokenText: EditText = view.findViewById(R.id.session_token_text)
 
-        var userId : String = ""
 
         sharedViewModel.isUserIdentified.observe(viewLifecycleOwner) { isUserIdentified ->
             isUserIdentifiedCheckbox.isChecked = isUserIdentified
@@ -49,8 +44,7 @@ class IdentificationFragment : Fragment() {
         identifyUnverifiedUserButton.setOnClickListener {
             if(unverifiedUserIdText.text.isNotBlank()) {
                 try {
-                    userId = unverifiedUserIdText.text.toString()
-                    viewModel.identifyUnverifiedUser(userId = userId)
+                    viewModel.identifyUnverifiedUser(userId = unverifiedUserIdText.text.toString())
                     alertDialogBox(getString(R.string.successful), getString(R.string.identification_successful))
                 } catch (e: Exception) {
                     alertDialogBox(getString(R.string.unsuccessful), getString(R.string.identification_unsuccessful))
@@ -61,8 +55,7 @@ class IdentificationFragment : Fragment() {
         identifyVerifiedUserButton.setOnClickListener {
             try {
                 if (verifiedUserIdText.text.isNotBlank() and sessionTokenText.text.isNotBlank()) {
-                    userId = verifiedUserIdText.text.toString()
-                    viewModel.identifyVerifiedUser(userId, sessionTokenText.text.toString())
+                    viewModel.identifyVerifiedUser(verifiedUserIdText.text.toString(), sessionTokenText.text.toString())
                     alertDialogBox(getString(R.string.successful), getString(R.string.identification_successful))
                 }
             }catch (e: Exception) {
@@ -72,21 +65,6 @@ class IdentificationFragment : Fragment() {
 
         logOutButton.setOnClickListener {
             viewModel.logout(requireContext())
-        }
-
-        updateUserButton.setOnClickListener {
-            try {
-                var userEmail: String? = null
-                if (updateUserEmail.text.isNotBlank()) {
-                    userEmail = updateUserEmail.text.toString()
-                }
-                if (userId.isNotEmpty()) {
-                    viewModel.updateUser(Identity(userId = userId, userInfo = UserInfo(userEmail)))
-                }
-                alertDialogBox(getString(R.string.successful), getString(R.string.update_user_successful))
-            } catch (e: Exception) {
-                alertDialogBox(getString(R.string.unsuccessful), getString(R.string.update_user_unsuccessful))
-            }
         }
     }