+++ sample/src/main/java/ai/devrev/sdk/sample/SessionAnalyticsFragment.kt
@@ -3,14 +3,13 @@ package ai.devrev.sdk.sample
 import ai.devrev.sdk.sample.viewmodel.SessionAnalyticsViewModel
 import ai.devrev.sdk.sample.viewmodel.SharedViewModel
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.CheckBox
-import android.widget.EditText
 import androidx.appcompat.app.AlertDialog
-import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 
 class SessionAnalyticsFragment : Fragment() {
@@ -28,10 +27,6 @@ class SessionAnalyticsFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        viewModel.trackScreen("Session Analytics")
-
-        sharedViewModel.setInScreenTransitioning(false)
-
         val isRecordingCheckbox: CheckBox = view.findViewById(R.id.is_session_recorded)
         val isMonitoringEnabledCheckBox: CheckBox = view.findViewById(R.id.is_monitoring_enabled)
         val startRecordingButton: Button  = view.findViewById(R.id.start_recording_button)
@@ -41,15 +36,6 @@ class SessionAnalyticsFragment : Fragment() {
         val processAllDemandsSessions: Button  = view.findViewById(R.id.process_all_on_demand_session_button)
         val stopAllMonitoring: Button  = view.findViewById(R.id.stop_all_monitoring_button)
         val resumeAllMonitoring: Button  = view.findViewById(R.id.resume_all_monitoring_button)
-        val startTimerWithPropertiesButton: Button = view.findViewById(R.id.start_timer_with_properties)
-        val endTimerWithPropertiesButton: Button = view.findViewById(R.id.end_timer_with_properties)
-        val unmaskedField: EditText = view.findViewById(R.id.manual_unmasked_item)
-
-        unmaskedField.tag = "devrev-unmask"
-
-        val properties = HashMap<String, Any>()
-        properties["state"] = "active"
-        viewModel.addSessionProperties(properties)
 
         sharedViewModel.isMonitoringEnabled.observe(viewLifecycleOwner) { isMonitoringEnabled ->
             isMonitoringEnabledCheckBox.isChecked = isMonitoringEnabled
@@ -119,30 +105,6 @@ class SessionAnalyticsFragment : Fragment() {
                 alertDialogBox(getString(R.string.monitoring_resumed), getString(R.string.monitoring_resume_success) )
             } catch (e: Exception) {
                 alertDialogBox(getString(R.string.monitoring_resume_failed), getString(R.string.monitoring_resume_error) )
-            }
-        }
-
-        startTimerWithPropertiesButton.setOnClickListener {
-            try {
-                val userProperties = HashMap<String, String>()
-                userProperties["key1"] = "value1"
-                userProperties["key2"] = "value2"
-                viewModel.startTimer("session", userProperties)
-                alertDialogBox(getString(R.string.start_timer_success), getString(R.string.timer_started))
-            } catch (e: Exception) {
-                alertDialogBox(getString(R.string.start_timer_error),getString(R.string.timer_not_started))
-            }
-        }
-
-        endTimerWithPropertiesButton.setOnClickListener {
-            try {
-                val userProperties = HashMap<String, String>()
-                userProperties["key3"] = "value3"
-                userProperties["key4"] = "value4"
-                viewModel.endTimer("session", userProperties)
-                alertDialogBox(getString(R.string.end_timer_success), getString(R.string.timer_ended))
-            } catch (e: Exception) {
-                alertDialogBox(getString(R.string.end_timer_error),getString(R.string.timer_not_ended))
             }
         }
     }