+++ sample/src/main/java/ai/devrev/sdk/sample/viewmodel/IdentificationViewModel.kt
@@ -18,10 +18,6 @@ class IdentificationViewModel(): ViewModel() {
         DevRev.identifyVerifiedUser(userId = userId, sessionToken = sessionToken)
     }
 
-    fun updateUser(identity: Identity) {
-        DevRev.updateUser(identity = identity)
-    }
-
     fun logout(context: Context) {
         val deviceId = DeviceInfo.getDeviceId(context)
         DevRev.logout(context, deviceId)